+++ src/network/utils/ipv4-address.cc
@@ -137,14 +137,13 @@ Ipv4Mask::GetPrefixLength() const
 }
 
 /**
- *  Value of a not-yet-initialized IPv4 address, corresponding to 102.102.102.102.
- *  This is totally arbitrary.
+ * Value of a not-yet-initialized IPv4 address, corresponding to 0.0.0.0.
+ * See also \RFC{1122}
  */
-static constexpr uint32_t UNINITIALIZED = 0x66666666U;
+static constexpr uint32_t UNINITIALIZED = 0x00000000U;
 
 Ipv4Address::Ipv4Address()
-    : m_address(UNINITIALIZED),
-      m_initialized(false)
+    : m_address(UNINITIALIZED)
 {
     NS_LOG_FUNCTION(this);
 }
@@ -153,7 +152,6 @@ Ipv4Address::Ipv4Address(uint32_t address)
 {
     NS_LOG_FUNCTION(this << address);
     m_address = address;
-    m_initialized = true;
 }
 
 Ipv4Address::Ipv4Address(const char* address)
@@ -162,13 +160,25 @@ Ipv4Address::Ipv4Address(const char* address)
 
     if (inet_pton(AF_INET, address, &m_address) <= 0)
     {
-        NS_LOG_LOGIC("Error, can not build an IPv4 address from an invalid string: " << address);
-        m_address = 0;
-        m_initialized = false;
+        NS_ABORT_MSG("Error, can not build an IPv4 address from an invalid string: " << address);
         return;
     }
-    m_initialized = true;
     m_address = ntohl(m_address);
+}
+
+bool
+Ipv4Address::CheckCompatibility(const std::string& addressStr)
+{
+    NS_LOG_FUNCTION(addressStr);
+
+    uint32_t buffer;
+
+    if (inet_pton(AF_INET, addressStr.c_str(), &buffer) <= 0)
+    {
+        NS_LOG_LOGIC("Error, can not build an IPv4 address from an invalid string: " << addressStr);
+        return false;
+    }
+    return true;
 }
 
 uint32_t
@@ -183,7 +193,6 @@ Ipv4Address::Set(uint32_t address)
 {
     NS_LOG_FUNCTION(this << address);
     m_address = address;
-    m_initialized = true;
 }
 
 void
@@ -192,12 +201,9 @@ Ipv4Address::Set(const char* address)
     NS_LOG_FUNCTION(this << address);
     if (inet_pton(AF_INET, address, &m_address) <= 0)
     {
-        NS_LOG_LOGIC("Error, can not build an IPv4 address from an invalid string: " << address);
-        m_address = 0;
-        m_initialized = false;
+        NS_ABORT_MSG("Error, can not build an IPv4 address from an invalid string: " << address);
         return;
     }
-    m_initialized = true;
     m_address = ntohl(m_address);
 }
 
@@ -237,7 +243,7 @@ bool
 Ipv4Address::IsInitialized() const
 {
     NS_LOG_FUNCTION(this);
-    return m_initialized;
+    return true;
 }
 
 bool
@@ -311,7 +317,6 @@ Ipv4Address::Deserialize(const uint8_t buf[4])
     ipv4.m_address |= buf[2];
     ipv4.m_address <<= 8;
     ipv4.m_address |= buf[3];
-    ipv4.m_initialized = true;
 
     return ipv4;
 }